+++ sdk/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to the @codebuff/sdk package will be documented in this file.
 
+## [0.1.8] - 2025-08-13
+
+### Added
+
+- `withAdditionalMessage` and `withMessageHistory` functions
+  - Add images, files, or other messages to a previous run
+  - Modify the history of any run
+- `initialSessionState` and `generateInitialRunState` functions
+  - Create a SessionState or RunState object from scratch
+
+### Removed
+
+- `getInitialSessionState` function
+
 ## [0.1.7] - 2025-08-12
 
 ### Updated types! AgentConfig has been renamed to AgentDefinition.
@@ -9,6 +23,7 @@ All notable changes to the @codebuff/sdk package will be documented in this file
 ## [0.1.5] - 2025-08-09
 
 ### Added
+
 - Complete `CodebuffClient`
 - Better docs
 - New `run()` api
@@ -16,6 +31,7 @@ All notable changes to the @codebuff/sdk package will be documented in this file
 ## [0.0.1] - 2025-08-05
 
 ### Added
+
 - Initial release of the Codebuff SDK
 - `CodebuffClient` class for interacting with Codebuff agents
 - `runNewChat` method for starting new chat sessions